Update incident service logs
This change modifies the service logs to more closely align with what we
currently have in WebRCA as well as address concerns with the incident
ID not making sense to external customers. More on that conversation
here:
https://redhat-internal.slack.com/archives/C03M8A471V1/p1704922462075899?thread_ts=1704899958.904489&cid=C03M8A471V1.

I also add an open-ended SL for sending updates to the customer. Often
this is handled by MCS in a separate proactive case, but I believe this would
still be useful.

OSDEV-1386